custom queries added in new repos

+++ server/src/main/java/com/ofg/hairdresser/repository/TreatmentRepository.java
@@ -3,5 +3,5 @@ package com.ofg.hairdresser.repository;
 import com.ofg.hairdresser.model.entity.Treatment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TreatmentRepository extends JpaRepository<Treatment, Integer> {
+public interface TreatmentRepository extends JpaRepository<Treatment, Long> {
 }
